Move the PaperWork into its own component
